Updating to modern redux using patterns recommended by redux team
Redux toolkit (RTK) handles a lot of the boilerplate for us now, so we don't need to include redux-thunk or some of the dev helpers. Unfortunately we are breaking some rules with our current implementation, so I had to turn off the immutable & serializable checks until we can investigate and refactor.
