+++ package.json
@@ -5,13 +5,13 @@
   "dependencies": {
     "@emotion/react": "11.4.1",
     "@emotion/styled": "11.3.0",
+    "@reduxjs/toolkit": "^1.6.2",
     "lodash.clonedeep": "4.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-keyboard-event-handler": "1.5.4",
     "react-redux": "7.2.5",
-    "react-router-dom": "^5.3.0",
-    "redux-thunk": "2.3.0"
+    "react-router-dom": "^5.3.0"
   },
   "devDependencies": {
     "@emotion/eslint-plugin": "^11.2.0",
